Simpler multi-platform support for pybind_extension targets.
Modified the pybind_extension() rule definition so that it creates a
target with the original name that switches between .pyd and .so
depending on the target platform. This platform-independent target is a
py_library(), so other py_library() targets can simply list it as a
dependency, the same way they would list other libraries.